requirements.txt
PyQt6>=6.4.0
PyYAML>=6.0
Jinja2>=3.0
